Break out GraphQL queries into their own files.
